requirements.txt
aiohttp==3.7.4.post0
astroid==2.9.3
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
colorama==0.4.4
discord==1.7.3
discord.py==1.7.3
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
multidict==6.0.2
platformdirs==2.5.1
pylint==2.12.2
python-dotenv==0.19.2
pytz==2021.3
toml==0.10.2
typing_extensions==4.1.1
wrapt==1.13.3
yarl==1.7.2